Style: 인터페이스 분리, 중복된 코드 삭제
