+++ app/_components/SideBar.tsx
@@ -57,12 +57,6 @@ export default function SideBar({ session, status }) {
                     className="btn btn-ghost"
                     onClick={() => toggleDropdown()}
                 >
-                <div
-                    tabIndex={0}
-                    role="button"
-                    className="btn btn-ghost"
-                    onClick={() => toggleDropdown()}
-                >
                     {status === 'authenticated' ? (
                         <Avatar user={session.user} />
                     ) : (
