+++ app/_interfaces/ButtonProps.ts
@@ -1,0 +1,8 @@
+interface DeletebuttonProps {
+    postId: number
+    commentId?: number
+    type: string
+    onDelete?: () => void
+}
+
+export type { DeletebuttonProps }
