+++ app/_interfaces/IColor.ts
@@ -1,0 +1,7 @@
+interface IColoredTagItem {
+    index: number
+    tagItem: string
+    deleteTagItem: (e: React.MouseEvent<HTMLButtonElement>) => void
+}
+
+export type { IColoredTagItem }
