+++ app/_interfaces/IComment.ts
@@ -1,0 +1,8 @@
+interface IComment {
+    id: number
+    post_id: number
+    parent_comment_id: number
+    author_id: number
+    content: string
+}
+export type { IComment }
